requirements.txt
pymysql
fastapi
pydantic
uvicorn
python-dotenv
cryptography
